Add cast for warning as error on Windows from r14201.
git-svn-id: http://skia.googlecode.com/svn/trunk@14202 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkFloatingPoint.h
@@ -40,7 +40,7 @@ static inline float sk_float_copysign(float x, float y) {
 
 // Visual studio prior to 13 only has 'double _copysign(double, double)'.
 #elif defined(_MSC_VER)
-    return _copysign(x, y);
+    return (float)_copysign(x, y);
 
 // Otherwise convert to bits and extract sign.
 #else
